# schema/openapi.yaml
components:
  schemas:
...
    ActivityRecord:
      additionalProperties: false
      properties:
        evidence:
          additionalProperties:
            type: string
          maxProperties: 32
          type: object
        id:
          maxLength: 64
          type: string
        identity:
          maxLength: 64
          type: string
        kind:
          $ref: '#/components/schemas/ActivityKind'
        message:
          maxLength: 512
          type: string
        occurredAt:
          format: date-time
          type: string
        result:
          $ref: '#/components/schemas/ActivityResult'
        target:
          maxLength: 128
          type: string
        targetType:
          enum:
          - host
          - unit
          - service
          - schedule
          - system
          type: string
      required:
      - id
      - occurredAt
      - kind
      - target
      - message
      type: object
    ActivityResponse:
      additionalProperties: false
      properties:
        records:
          items:
            $ref: '#/components/schemas/ActivityRecord'
          maxItems: 200
          type: array
        requestedAt:
          format: date-time
          type: string
        since:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
      required:
      - since
      - requestedAt
      - records
      type: object
...
    FilesystemEvidence:
      additionalProperties: false
      properties:
        device:
          maxLength: 240
          type: string
        freeBytes:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
        freeInodes:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
        fstype:
          maxLength: 64
          type: string
        inodeUsedPercent:
          anyOf:
          - maximum: 100.0
            minimum: 0.0
            type: number
          - type: 'null'
        mountPoint:
          maxLength: 240
          type: string
        state:
          $ref: '#/components/schemas/FilesystemState'
        totalBytes:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
        totalInodes:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
        usedBytes:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
        usedPercent:
          anyOf:
          - maximum: 100.0
            minimum: 0.0
            type: number
          - type: 'null'
      required:
      - mountPoint
      - device
      - fstype
      - totalBytes
      - freeBytes
      - usedBytes
      - usedPercent
      - totalInodes
      - freeInodes
      - inodeUsedPercent
      - state
      type: object
...
    HostCpuEvidence:
      additionalProperties: false
      properties:
        loadAverage15m:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
        loadAverage1m:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
        loadAverage5m:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
        utilizationPercent:
          anyOf:
          - maximum: 100.0
            minimum: 0.0
            type: number
          - type: 'null'
      required:
      - utilizationPercent
      - loadAverage1m
      - loadAverage5m
      - loadAverage15m
      type: object
    HostEvidence:
      additionalProperties: false
      properties:
        bootTime:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
        cpu:
          $ref: '#/components/schemas/HostCpuEvidence'
        filesystems:
          items:
            $ref: '#/components/schemas/FilesystemEvidence'
          type: array
        fresh:
          type: boolean
        hostname:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
        interfaces:
          items:
            $ref: '#/components/schemas/NetworkInterfaceEvidence'
          type: array
        memory:
          $ref: '#/components/schemas/HostMemoryEvidence'
        observedAt:
          format: date-time
          type: string
        pressure:
          items:
            $ref: '#/components/schemas/PressureEvidence'
          type: array
        providerMessage:
          anyOf:
          - maxLength: 240
            type: string
          - type: 'null'
        providerState:
          enum:
          - available
          - degraded
          - unavailable
          type: string
        temperatures:
          items:
            $ref: '#/components/schemas/TemperatureEvidence'
          type: array
        uptimeSeconds:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
      required:
      - observedAt
      - fresh
      - uptimeSeconds
      - bootTime
      - cpu
      - memory
      - filesystems
      - interfaces
      - pressure
      - temperatures
      - providerState
      type: object
    HostHistoryResponse:
      additionalProperties: false
      properties:
        observedAt:
          format: date-time
          type: string
        samples:
          items:
            $ref: '#/components/schemas/HostEvidence'
          type: array
        windowSeconds:
          minimum: 0.0
          type: integer
      required:
      - windowSeconds
      - observedAt
      - samples
      type: object
...
    ManagedScheduleDefinition:
      additionalProperties: false
      properties:
        arguments:
          items:
            type: string
          maxItems: 64
          type: array
        description:
          maxLength: 256
          type: string
        executable:
          maxLength: 256
          minLength: 1
          pattern: ^/[^ \t]+$
          type: string
        missedRunBehavior:
          enum:
          - ignore
          - catch_up
          type: string
        name:
          maxLength: 64
          minLength: 1
          pattern: ^[a-z][a-z0-9-]*$
          type: string
        onCalendar:
          maxLength: 128
          minLength: 1
          type: string
        overlapPolicy:
          enum:
          - allow
          - drop
          - queue
          type: string
        relatedService:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
        timeoutSeconds:
          maximum: 86400.0
          minimum: 1.0
          type: integer
        user:
          maxLength: 64
          pattern: ^[a-z_][a-z0-9_-]*$
          type: string
        workingDirectory:
          anyOf:
          - maxLength: 256
            pattern: ^/[^ \t]*$
            type: string
          - type: 'null'
      required:
      - name
      - onCalendar
      - executable
      type: object
    ManagedScheduleView:
      additionalProperties: false
      properties:
        definition:
          $ref: '#/components/schemas/ManagedScheduleDefinition'
        enabled:
          type: boolean
        installed:
          type: boolean
        lastResult:
          $ref: '#/components/schemas/ScheduleResult'
        lastRun:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
        revision:
          minimum: 0.0
          type: integer
        serviceUnit:
          maxLength: 128
          type: string
        timerUnit:
          maxLength: 128
          type: string
      required:
      - definition
      - revision
      - serviceUnit
      - timerUnit
      - lastRun
      type: object
    NetworkInterfaceEvidence:
      additionalProperties: false
      properties:
        name:
          maxLength: 64
          type: string
        rxBytesPerSecond:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
        rxBytesTotal:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
        state:
          $ref: '#/components/schemas/InterfaceState'
        txBytesPerSecond:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
        txBytesTotal:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
      required:
      - name
      - state
      - rxBytesTotal
      - txBytesTotal
      - rxBytesPerSecond
      - txBytesPerSecond
      type: object
...
    OperationRequest:
      additionalProperties: false
      properties:
        expectedState:
          anyOf:
          - $ref: '#/components/schemas/UnitState'
          - type: 'null'
        operation:
          $ref: '#/components/schemas/OperationKind'
        reason:
          maxLength: 240
          type: string
      required:
      - operation
      type: object
    OperationResult:
      additionalProperties: false
      properties:
        evidence:
          additionalProperties:
            type: string
          maxProperties: 32
          type: object
        id:
          maxLength: 64
          type: string
        message:
          maxLength: 512
          type: string
        operation:
          $ref: '#/components/schemas/OperationKind'
        reconciledAt:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
        requestedAt:
          format: date-time
          type: string
        state:
          $ref: '#/components/schemas/OperationState'
        unit:
          maxLength: 128
          type: string
      required:
      - id
      - unit
      - operation
      - state
      - requestedAt
      - reconciledAt
      type: object
...
    ScheduleInventoryResponse:
      additionalProperties: false
      properties:
        fresh:
          type: boolean
        generatedAt:
          format: date-time
          type: string
        schedules:
          items:
            $ref: '#/components/schemas/ScheduleSnapshot'
          type: array
      required:
      - generatedAt
      - fresh
      - schedules
      type: object
    ScheduleResult:
      enum:
      - success
      - failed
      - not_observed
      - pending
      - unknown
      type: string
    ScheduleSnapshot:
      additionalProperties: false
      properties:
        enabled:
          type: boolean
        humanReadable:
          maxLength: 256
          type: string
        id:
          maxLength: 128
          type: string
        lastResult:
          $ref: '#/components/schemas/ScheduleResult'
        lastRun:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
        managed:
          type: boolean
        name:
          maxLength: 128
          type: string
        nextRun:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
        owner:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
        provenance:
          maxLength: 256
          type: string
        rawExpression:
          maxLength: 512
          type: string
        relatedService:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
        source:
          $ref: '#/components/schemas/ScheduleSource'
        target:
          maxLength: 256
          type: string
      required:
      - id
      - name
      - source
      - nextRun
      - lastRun
      type: object
...
    TemperatureEvidence:
      additionalProperties: false
      properties:
        celsius:
          anyOf:
          - type: number
          - type: 'null'
        zone:
          maxLength: 128
          type: string
      required:
      - zone
      - celsius
      type: object
...
    UnitInventoryResponse:
      additionalProperties: false
      properties:
        counts:
          additionalProperties:
            type: integer
          description: Counts by active state.
          type: object
        fresh:
          type: boolean
        generatedAt:
          format: date-time
          type: string
        units:
          items:
            $ref: '#/components/schemas/UnitSnapshot'
          type: array
      required:
      - generatedAt
      - fresh
      - counts
      - units
      type: object
    UnitLogsResponse:
      additionalProperties: false
      properties:
        records:
          items:
            $ref: '#/components/schemas/LogRecord'
          maxItems: 500
          type: array
        requestedAt:
          format: date-time
          type: string
        truncated:
          type: boolean
        unit:
          maxLength: 128
          type: string
      required:
      - unit
      - requestedAt
      - records
      - truncated
      type: object
...
  securitySchemes:
    authentikIdentity:
      in: header
      name: X-Authentik-Username
      type: apiKey
info:
  description: Read-only API for the single-VPS Hyperion service atlas.
  version: 1.0.0
openapi: 3.1.0
paths:
  /api/v1/activity:
    get:
      description: Get recorded activity events, newest-first.
      operationId: getActivity
      parameters:
      - in: query
        name: since
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
      - in: query
        name: limit
        required: false
        schema:
          maximum: 200
          minimum: 1
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
          description: Recorded events newest-first.
        '401':
          description: Authentik identity header is absent.
      security:
      - APIKeyHeader: []
      summary: Get recorded activity
  /api/v1/host:
    get:
      description: Get the current bounded host evidence sample.
      operationId: getHost
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostEvidence'
          description: Current host metric evidence.
        '401':
          description: Authentik identity header is absent.
      security:
      - APIKeyHeader: []
      summary: Get current host evidence
  /api/v1/host/history:
    get:
      description: Get bounded recent host samples within the requested window.
      operationId: getHostHistory
      parameters:
      - in: query
        name: window
        required: false
        schema:
          pattern: ^\d+[smh]$
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostHistoryResponse'
          description: Bounded recent host samples, oldest first.
        '401':
          description: Authentik identity header is absent.
      security:
      - APIKeyHeader: []
      summary: Get recent host history samples
  /api/v1/schedules:
    get:
      description: Get the unified schedule inventory from systemd timers and cron.
      operationId: getSchedules
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleInventoryResponse'
          description: Normalized systemd timer and cron schedule index.
        '401':
          description: Authentik identity header is absent.
      security:
      - APIKeyHeader: []
      summary: Get the schedule inventory
    post:
      description: Create a managed schedule as a systemd service/timer pair.
      operationId: createSchedule
      parameters:
      - in: header
        name: Origin
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagedScheduleDefinition'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedScheduleView'
          description: The installed managed schedule.
        '400':
          description: Definition failed validation.
        '401':
          description: Authentik identity header is absent.
      security:
      - APIKeyHeader: []
      summary: Create a Hyperion-managed schedule
  /api/v1/schedules/{scheduleId}:
    delete:
      description: Delete a managed schedule; the timer and service units are removed.
      operationId: deleteSchedule
      parameters:
      - in: path
        name: scheduleId
        required: true
        schema:
          type: string
      - in: query
        name: expectedRevision
        required: false
        schema:
          minimum: 0
          type: integer
      - in: header
        name: Origin
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
      responses:
        '204':
          description: The managed schedule was deleted.
        '400':
          description: Revision or existence validation failed.
        '401':
          description: Authentik identity header is absent.
      security:
      - APIKeyHeader: []
      summary: Delete a Hyperion-managed schedule
    get:
      description: Get one schedule record by inventory identity.
      operationId: getSchedule
      parameters:
      - in: path
        name: scheduleId
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleSnapshot'
          description: A single schedule record.
        '401':
          description: Authentik identity header is absent.
        '404':
          description: Schedule is not in the inventory.
      security:
      - APIKeyHeader: []
      summary: Get one schedule record
    put:
      description: Update a managed schedule with optimistic-concurrency revision
        control.
      operationId: updateSchedule
      parameters:
      - in: path
        name: scheduleId
        required: true
        schema:
          type: string
      - in: query
        name: expectedRevision
        required: false
        schema:
          minimum: 0
          type: integer
      - in: header
        name: Origin
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagedScheduleDefinition'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedScheduleView'
          description: The updated managed schedule.
        '400':
          description: Definition or revision failed validation.
        '401':
          description: Authentik identity header is absent.
      security:
      - APIKeyHeader: []
      summary: Update a Hyperion-managed schedule
...
  /api/v1/units:
    get:
      description: Get the systemd service inventory with per-unit property evidence.
      operationId: getUnits
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitInventoryResponse'
          description: Curated systemd service inventory.
        '401':
          description: Authentik identity header is absent.
      security:
      - APIKeyHeader: []
      summary: Get the systemd unit inventory
  /api/v1/units/{unitName}/logs:
    get:
      description: Get bounded journal logs for one server-enumerated unit.
      operationId: getUnitLogs
      parameters:
      - in: path
        name: unitName
        required: true
        schema:
          type: string
      - in: query
        name: tail
        required: false
        schema:
          enum:
          - 50
          - 100
          - 250
          - 500
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitLogsResponse'
          description: Journal logs for the unit, oldest to newest.
        '401':
          description: Authentik identity header is absent.
        '404':
          description: Unit is not in the server-enumerated inventory.
        '503':
          description: Journal is temporarily unavailable.
      security:
      - APIKeyHeader: []
      summary: Get bounded journal logs for one unit
  /api/v1/units/{unitName}/operations:
    post:
      description: Request one typed operation; idempotent per idempotency key.
      operationId: postOperation
      parameters:
      - in: path
        name: unitName
        required: true
        schema:
          type: string
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 64
          minLength: 8
          type: string
      - in: header
        name: Origin
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResult'
          description: The operation result, pending or terminal.
        '401':
          description: Authentik identity header is absent.
      security:
      - APIKeyHeader: []
      summary: Request a typed systemd operation
...
security:
- authentikIdentity: []
servers:
- url: /